wal: add thread shutdown triggers and destroy writer via trigger
Because of obuf, an xlog may not be used from different threads, so any
xlog written by the WAL thread must be closed from the WAL thread on
shutdown. The problem is, on shutdown event loop is already stopped, so
we can't send a message via cbus. That's why we close the write-ahead
log explicitly from the WAL thread routine. This looks fine as long as
there's the only log written by the WAL thread, write-ahead log.
However, this is about to change - we are going to write vinyl metadata
log via the WAL thread too. Closing the vinyl metadata log explicitly
from the thread routine would violate encapsulation, so let's introduce
shutdown triggers called by the WAL thread before exiting.