vinyl: write metadata log via wal thread
We can't write to xlog from a coeio task, because coeio uses a pool of
workers while xlog can only be used from a single thread due to obuf.
As a result, we are bound to use various hacks, like reopening xlog on
rotation or encoding log records in tx. The hacks are getting more and
more ugly, nearly unbearable, especially when it comes to separating
garbage collection from log rotation. So let's write the metadata log on
behalf of the wal thread.

Notes:
 - engine_init() -> vy_env_new() calls vy_log_new(), which needs
   the wal thread initialized to register a shutdown trigger, so
   wal_thread_start() is moved above engine_init() in box_cfg().
 - garbage collection is now called from the tx thread, because it
   involves communication with wal, so use coeio_unlink() instead
   of unlink() in vy_run_unlink_files().