vinyl: log drop index only when last ref is gone
When a vinyl index is dropped, there still might be dump/compaction
tasks running for it. Upon completion a task will log run/range creation
or deletion, which will cause recovery to fail if the index is dropped.
To avoid this, let's write drop index record to the metadata log only
when the last reference to the index has gone.