vinyl: separate gc from metadata log rotation
This patch is a preparation for centralized garbage collection. It
extracts the code doing garbage collection from vy_log_rotate() and
places it in a separate function vy_log_collect_garbage(). Apart from a
vy_log object, the latter takes a signature which determines the minimal
age an object has to have to be deleted - the function only deletes runs
that were deleted before the log received the given signature. It is
needed for making vinyl respect box.cfg.snapshot_count.