vinyl: delete old vymeta files on gc
To keep track of rotated log files, add two new record types,
VY_LOG_ROTATE and VY_LOG_FORGET. The former is written when the
metadata log is rotated and contains the signature of the old log.
The latter is written when an old log file is deleted and contains
the signature of the deleted log.